+++ packages/template-retail-react-app/CHANGELOG.md
@@ -1,5 +1,10 @@
 ## v2.5.0-dev (Apr 17, 2024)
+
 - Use `stale-while-revalidate` cache control directive [#1744](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/1744)
+
+### Bug Fixes
+
+-   Fix seo component not settings keywords meta tag [#1762](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/1762)
 
 ## v2.4.1 (Apr 17, 2024)
 - Update dependency commerce-sdk-react to 1.4.1 to have access to all params keys for Shopper Search [#1750](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/1750)